-changed names of the custom components to be consistent by starting with the prefix MSTeams -removed unnecessary Layout components -small refactoring
