+++ docs/src/prototypes/meetingOptions/components/MSTeamsFooter.tsx
