+++ docs/src/prototypes/meetingOptions/components/MSTeamsHeader.tsx
@@ -1,0 +1,17 @@
+import * as React from 'react'
+import { Provider } from '@stardust-ui/react'
+import { middleColumnStyles } from '../styles'
+
+export default props => {
+  return (
+    <Provider.Consumer
+      render={({ siteVariables }) => {
+        return (
+          <div style={{ backgroundColor: siteVariables.brand }}>
+            <div style={{ ...middleColumnStyles }}>{props.content}</div>
+          </div>
+        )
+      }}
+    />
+  )
+}
