+++ docs/src/prototypes/meetingOptions/components/MSTeamsLink.tsx
@@ -1,0 +1,15 @@
+import * as React from 'react'
+import { Provider, Text } from '@stardust-ui/react'
+
+export default props => {
+  const { content, children } = props
+  return (
+    <Provider.Consumer
+      render={({ siteVariables }) => (
+        <Text as="a" content={content} styles={{ color: siteVariables.brand }}>
+          {children}
+        </Text>
+      )}
+    />
+  )
+}
