+++ docs/src/prototypes/meetingOptions/index.tsx
@@ -1,19 +1,29 @@
 import * as React from 'react'
-import Header from './components/Header'
-import Footer from './components/Footer'
+import MSTeamsHeader from './components/MSTeamsHeader'
+import MSTeamsFooter from './components/MSTeamsFooter'
 import { mainStyle } from './styles'
-import Logo from './components/Logo'
+import MSTeamsLogo from './components/MSTeamsLogo'
 import MSTeamsDivider from './components/MSTeamsDivider'
-import MeetingOptions from './components/MeetingOptions'
+import MSTeamsMeetingOptions from './components/MSTeamsMeetingOptions'
+import MSTeamsLink from './components/MSTeamsLink'
 
 class MeetingOptionsPrototype extends React.Component<any, any> {
   render() {
     return (
       <div style={mainStyle}>
-        <Header content={<Logo icon={{ name: 'team-create' }} companyName="Microsoft Teams" />} />
+        <MSTeamsHeader
+          content={<MSTeamsLogo icon={{ name: 'team-create' }} companyName="Microsoft Teams" />}
+        />
         <MSTeamsDivider transparent size={40} />
-        <MeetingOptions />
-        <Footer content="&copy;Microsoft Corporation. All Rights reserved | Legal Notice | Privacy Policy" />
+        <MSTeamsMeetingOptions />
+        <MSTeamsFooter
+          content={
+            <span>
+              &copy;Microsoft Corporation. All Rights reserved |{' '}
+              <MSTeamsLink content="Legal Notice" /> | <MSTeamsLink content="Privacy Policy" />
+            </span>
+          }
+        />
       </div>
     )
   }
